# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estuaire Emploie API Documentation",
        "description": "Documentation de l'API REST pour l'application mobile Estuaire Emploie - Plateforme de recherche d'emploi au Cameroun",
        "contact": {
            "email": "contact@estuaire-emploie.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API Principal"
        }
    ],
    "paths": {
        "/api/jobs/{id}/apply": {
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Postuler à une offre d'emploi",
                "description": "Permet à un candidat de postuler à une offre avec CV, lettre de motivation et portfolio",
                "operationId": "40e59a18fb7f7e8eaef58d6258674856",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "cv": {
                                        "description": "Fichier CV (PDF, DOC, DOCX - max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "cover_letter": {
                                        "description": "Lettre de motivation",
                                        "type": "string"
                                    },
                                    "portfolio_url": {
                                        "description": "URL du portfolio",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidature soumise avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Candidature soumise avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Vous avez déjà postulé à cette offre"
                    },
                    "404": {
                        "description": "Offre non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation (CV requis, format invalide, etc.)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/my-applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Mes candidatures",
                "operationId": "140e8822ac97f9afa742a5d5e5595d77",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "viewed",
                                "shortlisted",
                                "rejected",
                                "interview",
                                "accepted"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste de mes candidatures",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/applications/{id}": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Détails d'une candidature",
                "operationId": "1068983793babb1ce7b105cee2aeade4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la candidature",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Candidature non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recruiter/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Candidatures reçues (Recruteur)",
                "description": "Récupère toutes les candidatures reçues pour les offres d'emploi de l'entreprise du recruteur connecté",
                "operationId": "getReceivedApplications",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "viewed",
                                "shortlisted",
                                "rejected",
                                "interview",
                                "accepted"
                            ]
                        }
                    },
                    {
                        "name": "job_id",
                        "in": "query",
                        "description": "Filtrer par offre d'emploi",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des candidatures reçues",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Vous n'êtes pas recruteur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vous n'êtes pas recruteur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/applications/{id}/status": {
            "patch": {
                "tags": [
                    "Applications"
                ],
                "summary": "Mettre à jour le statut d'une candidature (Recruteur)",
                "description": "Permet au recruteur de changer le statut d'une candidature (acceptée, refusée, en révision, etc.)",
                "operationId": "updateApplicationStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "viewed",
                                            "shortlisted",
                                            "rejected",
                                            "interview",
                                            "accepted"
                                        ],
                                        "example": "viewed"
                                    },
                                    "internal_notes": {
                                        "type": "string",
                                        "example": "Bon profil, à convoquer pour entretien"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Statut de la candidature mis à jour"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Non autorisé - Cette candidature ne concerne pas votre entreprise"
                    },
                    "404": {
                        "description": "Candidature non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum - Utiliser le token obtenu lors du login/register",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}